+++ internal/resources/domaincertificatesettings/domain_certificate_settings_model.go
@@ -1,0 +1,9 @@
+package domaincertificatesettings
+
+import "github.com/hashicorp/terraform-plugin-framework/types"
+
+type DomainCertificateSettingsModel struct {
+	ID                    types.String `tfsdk:"id"`
+	DomainID              types.String `tfsdk:"domain_id"`
+	FallbackCertificateID types.String `tfsdk:"fallback_certificate_id"`
+}